+++ src/helpers/info.js
@@ -56,7 +56,7 @@ export const rename = `
 export const flags = `
    |
    |  flags:
-   |     ${chalk.yellow('-h, --help')} : get command's information
+   |     ${chalk.yellow('-h, --help')} : get command information
    |     ${chalk.yellow('-v, --version')} : get cithak version
    |     ${chalk.yellow('-p, --path')} : add extra path relative to current working directory or a git repo
    |     ${chalk.gray('( ex: -p ./my-folder or -p github@username/repo )')}
@@ -65,9 +65,11 @@ export const flags = `
    |     ${chalk.yellow('-s, --safe')} : copy template without replacing existing files
    |     ${chalk.yellow('-c, --change')} : change file that will be executed
    |     ${chalk.gray('( ex: \'npm|pnpm\' ) : npm i package -> pnpm i package')}
-
+   
+   |     ${chalk.yellow('-i, --index')} : prefer index and not overwrite files
    |     ${chalk.yellow('--nj, --no-join')} : copy template without join joinable data file like json file
    |     ${chalk.yellow('--ne, --no-exec')} : copy template without execute any tasks
+   |     ${chalk.yellow('--sl, --silent')} : hide activity detail logs
 `
 
 export const complete =